Fix existing compiler warnings

+++ prl-tools-lin/kmods/prl_tg/Toolgate/Guest/Linux/prl_tg/prltg_common.h
@@ -151,7 +151,7 @@ tg_in32(struct tg_dev *dev, unsigned long port)
 #if defined(__aarch64__)
 	x = ioread32(dev->base_addr + port);
 #else
-	x = inl(dev->base_addr + port);
+	x = inl((unsigned long) dev->base_addr + port);
 #endif
 	spin_unlock_irqrestore(&dev->lock, flags);
 	return (x);
@@ -166,7 +166,7 @@ tg_out32(struct tg_dev *dev, unsigned long port, u32 val)
 #if defined(__aarch64__)
 	iowrite32(val, dev->base_addr + port);
 #else
-	outl(val, dev->base_addr + port);
+	outl(val, (unsigned long) dev->base_addr + port);
 #endif
 	spin_unlock_irqrestore(&dev->lock, flags);
 }
@@ -176,7 +176,7 @@ tg_out(struct tg_dev *dev, unsigned long port, unsigned long long val)
 {
 	unsigned long flags;
 
-	port += dev->base_addr;
+	port += (unsigned long) dev->base_addr;
 	spin_lock_irqsave(&dev->lock, flags);
 
 #if defined(__aarch64__)
